gif-gen/requirements.txt
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.3.0
